first work at refactoring queue into seprate folder
