+++ src/server/server.ts
@@ -6,8 +6,8 @@ import * as path from "path"
 import * as Raven from "raven"
 import * as winston from "winston"
 import * as Hub from "../hub"
+import * as ExecuteProcessQueue from "../xpc/execute_process_queue"
 import * as apiKey from "./api_key"
-import * as ExecuteProcessQueue from "./execute_process_queue"
 
 const expressWinston = require("express-winston")
 const blocked = require("blocked-at")
